+++ test/Reflection/ReflectionFunctionTest.php
@@ -1,27 +1,29 @@
 <?php
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/zendframework/zend-server for the canonical source repository
+ * @copyright Copyright (c) 2005-2018 Zend Technologies USA Inc. (https://www.zend.com)
+ * @license   https://github.com/zendframework/zend-server/blob/master/LICENSE.md New BSD License
  */
 
 namespace ZendTest\Server\Reflection;
 
+use PHPUnit\Framework\TestCase;
+use ReflectionFunction;
 use Zend\Server\Reflection;
+use Zend\Server\Reflection\AbstractFunction;
+use Zend\Server\Reflection\Prototype;
 
 /**
  * @group      Zend_Server
  */
-class ReflectionFunctionTest extends \PHPUnit_Framework_TestCase
+class ReflectionFunctionTest extends TestCase
 {
     public function testConstructor()
     {
-        $function = new \ReflectionFunction('\ZendTest\Server\Reflection\function1');
+        $function = new ReflectionFunction('\ZendTest\Server\Reflection\function1');
         $r = new Reflection\ReflectionFunction($function);
-        $this->assertInstanceOf('Zend\Server\Reflection\ReflectionFunction', $r);
-        $this->assertInstanceOf('Zend\Server\Reflection\AbstractFunction', $r);
+        $this->assertInstanceOf(\Zend\Server\Reflection\ReflectionFunction::class, $r);
+        $this->assertInstanceOf(AbstractFunction::class, $r);
         $params = $r->getParameters();
 
         $r = new Reflection\ReflectionFunction($function, 'namespace');
@@ -39,7 +41,7 @@ class ReflectionFunctionTest extends \PHPUnit_Framework_TestCase
 
     public function testPropertyOverloading()
     {
-        $function = new \ReflectionFunction('\ZendTest\Server\Reflection\function1');
+        $function = new ReflectionFunction('\ZendTest\Server\Reflection\function1');
         $r = new Reflection\ReflectionFunction($function);
 
         $r->system = true;
@@ -49,7 +51,7 @@ class ReflectionFunctionTest extends \PHPUnit_Framework_TestCase
 
     public function testNamespace()
     {
-        $function = new \ReflectionFunction('\ZendTest\Server\Reflection\function1');
+        $function = new ReflectionFunction('\ZendTest\Server\Reflection\function1');
         $r = new Reflection\ReflectionFunction($function, 'namespace');
         $this->assertEquals('namespace', $r->getNamespace());
         $r->setNamespace('framework');
@@ -58,7 +60,7 @@ class ReflectionFunctionTest extends \PHPUnit_Framework_TestCase
 
     public function testDescription()
     {
-        $function = new \ReflectionFunction('\ZendTest\Server\Reflection\function1');
+        $function = new ReflectionFunction('\ZendTest\Server\Reflection\function1');
         $r = new Reflection\ReflectionFunction($function);
         $this->assertContains('function for reflection', $r->getDescription());
         $r->setDescription('Testing setting descriptions');
@@ -67,41 +69,41 @@ class ReflectionFunctionTest extends \PHPUnit_Framework_TestCase
 
     public function testGetPrototypes()
     {
-        $function = new \ReflectionFunction('\ZendTest\Server\Reflection\function1');
+        $function = new ReflectionFunction('\ZendTest\Server\Reflection\function1');
         $r = new Reflection\ReflectionFunction($function);
 
         $prototypes = $r->getPrototypes();
         $this->assertInternalType('array', $prototypes);
-        $this->assertEquals(8, count($prototypes));
+        $this->assertCount(8, $prototypes);
 
         foreach ($prototypes as $p) {
-            $this->assertInstanceOf('Zend\Server\Reflection\Prototype', $p);
+            $this->assertInstanceOf(Prototype::class, $p);
         }
     }
 
     public function testGetPrototypes2()
     {
-        $function = new \ReflectionFunction('\ZendTest\Server\Reflection\function2');
+        $function = new ReflectionFunction('\ZendTest\Server\Reflection\function2');
         $r = new Reflection\ReflectionFunction($function);
 
         $prototypes = $r->getPrototypes();
         $this->assertInternalType('array', $prototypes);
         $this->assertNotEmpty($prototypes);
-        $this->assertEquals(1, count($prototypes));
+        $this->assertCount(1, $prototypes);
 
         foreach ($prototypes as $p) {
-            $this->assertInstanceOf('Zend\Server\Reflection\Prototype', $p);
+            $this->assertInstanceOf(Prototype::class, $p);
         }
     }
 
 
     public function testGetInvokeArguments()
     {
-        $function = new \ReflectionFunction('\ZendTest\Server\Reflection\function1');
+        $function = new ReflectionFunction('\ZendTest\Server\Reflection\function1');
         $r = new Reflection\ReflectionFunction($function);
         $args = $r->getInvokeArguments();
         $this->assertInternalType('array', $args);
-        $this->assertEquals(0, count($args));
+        $this->assertCount(0, $args);
 
         $argv = ['string1', 'string2'];
         $r = new Reflection\ReflectionFunction($function, null, $argv);
@@ -111,28 +113,28 @@ class ReflectionFunctionTest extends \PHPUnit_Framework_TestCase
 
     public function testClassWakeup()
     {
-        $function = new \ReflectionFunction('\ZendTest\Server\Reflection\function1');
+        $function = new ReflectionFunction('\ZendTest\Server\Reflection\function1');
         $r = new Reflection\ReflectionFunction($function);
         $s = serialize($r);
         $u = unserialize($s);
-        $this->assertInstanceOf('Zend\Server\Reflection\ReflectionFunction', $u);
+        $this->assertInstanceOf(\Zend\Server\Reflection\ReflectionFunction::class, $u);
         $this->assertEquals('', $u->getNamespace());
     }
 
     public function testMultipleWhitespaceBetweenDoctagsAndTypes()
     {
-        $function = new \ReflectionFunction('\ZendTest\Server\Reflection\function3');
+        $function = new ReflectionFunction('\ZendTest\Server\Reflection\function3');
         $r = new Reflection\ReflectionFunction($function);
 
         $prototypes = $r->getPrototypes();
         $this->assertInternalType('array', $prototypes);
         $this->assertNotEmpty($prototypes);
-        $this->assertEquals(1, count($prototypes));
+        $this->assertCount(1, $prototypes);
 
         $proto = $prototypes[0];
         $params = $proto->getParameters();
         $this->assertInternalType('array', $params);
-        $this->assertEquals(1, count($params));
+        $this->assertCount(1, $params);
         $this->assertEquals('string', $params[0]->getType());
     }
 
@@ -141,7 +143,7 @@ class ReflectionFunctionTest extends \PHPUnit_Framework_TestCase
      */
     public function testParameterReflectionShouldReturnTypeAndVarnameAndDescription()
     {
-        $function = new \ReflectionFunction('\ZendTest\Server\Reflection\function1');
+        $function = new ReflectionFunction('\ZendTest\Server\Reflection\function1');
         $r = new Reflection\ReflectionFunction($function);
 
         $prototypes = $r->getPrototypes();