+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
-## 2.7.1 - TBD
+## 2.8.0 - 2018-04-30
 
 ### Added
+
+- [#26](https://github.com/zendframework/zend-server/pull/26) adds support for PHP 7.1 and 7.2.
+
+- [#19](https://github.com/zendframework/zend-server/pull/19) adds the ability to register any PHP callable with `Zend\Server\Method\Callback`.
+
+### Changed
 
 - Nothing.
 
@@ -14,11 +20,13 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Removed
 
-- Nothing.
+- [#26](https://github.com/zendframework/zend-server/pull/26) removes support for HHVM.
 
 ### Fixed
 
-- Nothing.
+- [#20](https://github.com/zendframework/zend-server/pull/20) fixes how `Cache::save()` works when `Server::getFunctions()` returns an
+  associative array instead of a `Definition`, ensuring it will also skip
+  any blacklisted methods when used in this way.
 
 ## 2.7.0 - 2016-06-20